api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
alembic>=1.14.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
screamingfrog>=0.2.2
ruff>=0.8.0
pytest>=8.0.0
google-auth>=2.0.0
google-cloud-compute>=1.19.0
google-cloud-storage>=2.18.0
croniter>=2.0.0
